Enhance BOM Plugin
* more styling in the Options records
* done for (some) edges
* still to be done for nodes
