+++ src/AasxPluginBomStructure/AasxPluginBomStructure.Options.json
@@ -1,4 +1,27 @@
-﻿{
+﻿/*
+  Help on LinkStyles
+  ==================
+
+  "LinkStyles": [
+        {
+          "Match": {
+            "type": "ConceptDescription",
+            "local": false,
+            "value": "http://admin-shell.io/sandbox/CompositeComponent/General/IsPartOfForBOM/1/0",
+            "index": 0,
+            "idType": "IRI"
+          },
+          "Direction": "Forward" | "Backward" | "None" | "Both",
+          "Color": "DarkGreen" | "#ff223344",
+          "Width": 3.0,
+          "Skip": false,       // Skip this link, do not draw it
+          "Dashed": false,
+          "Bold": false,
+          "Dotted": false
+        }
+      ]
+*/
+{
   "Records": [
     {
       "AllowSubmodelSemanticId": [
@@ -30,6 +53,23 @@
           "value": "http://admin-shell.io/sandbox/CompositeComponent/General/SubmodelBOM/1/0",
           "index": 0,
           "idType": "IRI"
+        }
+      ],
+      "Layout": 3,
+      "Compact":  true,
+      "LinkStyles": [
+        {
+          "Match": {
+            "type": "ConceptDescription",
+            "local": false,
+            "value": "http://admin-shell.io/sandbox/CompositeComponent/General/IsPartOfForBOM/1/0",
+            "index": 0,
+            "idType": "IRI"
+          },
+          "Direction": "Forward",
+          "Color": "DarkGreen",
+          "Width": 1.0,
+          "Text":  "isPartOf"
         }
       ]
     }
