+++ src/AasxPluginBomStructure/BomStructureOptions.cs
@@ -13,12 +13,75 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using AdminShellNS;
+using Newtonsoft.Json;
 
 namespace AasxPluginBomStructure
 {
+    public enum BomLinkDirection { None, Forward, Backward, Both }
+
+    public class BomLinkStyle : AdminShell.IGetReference
+    {
+        public AdminShell.Key Match;
+        public bool Skip;
+        public BomLinkDirection Direction;
+        public string Color;
+        public double Width;
+        public bool Dashed, Bold, Dotted;
+        public string Text;
+
+        public AdminShell.Reference GetReference() { return new AdminShell.Reference(Match); }
+    }
+
+    public class BomLinkStyleList : List<BomLinkStyle>
+    {
+        [JsonIgnore]
+        public AasReferenceStore<BomLinkStyle> Store = new AasReferenceStore<BomLinkStyle>();
+
+        public void Index()
+        {
+            foreach (var ls in this)
+                Store.Index(AdminShell.Reference.CreateNew(ls.Match), ls);
+        }
+    }
+
     public class BomStructureOptionsRecord
     {
         public List<AdminShell.Key> AllowSubmodelSemanticId = new List<AdminShell.Key>();
+
+        public int Layout;
+        public bool? Compact;
+
+        public BomLinkStyleList LinkStyles = new BomLinkStyleList();
+
+        public void Index()
+        {
+            LinkStyles.Index();
+        }
+
+        public BomLinkStyle FindFirstLinkStyle(AdminShell.SemanticId semId)
+        {
+            if (semId == null)
+                return null;
+            return LinkStyles.Store.FindElementByReference(semId, AdminShell.Key.MatchMode.Relaxed);
+        }
+    }
+
+    public class BomStructureOptionsRecordList : List<BomStructureOptionsRecord>
+    {
+        public BomStructureOptionsRecordList() { }
+
+        public BomStructureOptionsRecordList(IEnumerable<BomStructureOptionsRecord> collection) : base(collection) { }
+
+        public BomLinkStyle FindFirstLinkStyle(AdminShell.SemanticId semId)
+        {
+            foreach (var rec in this)
+            {
+                var res = rec.FindFirstLinkStyle(semId);
+                if (res != null)
+                    return res;
+            }
+            return null;
+        }
     }
 
     public class BomStructureOptions : AasxIntegrationBase.AasxPluginOptionsBase
@@ -41,6 +104,27 @@ namespace AasxPluginBomStructure
             opt.Records.Add(rec);
 
             return opt;
+        }
+
+        /// <summary>
+        /// For faster access of styles, index them by a hash map
+        /// </summary>
+        public void Index()
+        {
+            foreach (var rec in Records)
+                rec.Index();
+        }
+
+        /// <summary>
+        /// Find matching options records
+        /// </summary>
+        public IEnumerable<BomStructureOptionsRecord> MatchingRecords(AdminShell.SemanticId semId)
+        {
+            foreach (var rec in Records)
+                if (rec.AllowSubmodelSemanticId != null)
+                    foreach (var x in rec.AllowSubmodelSemanticId)
+                        if (semId != null && semId.Matches(x))
+                            yield return rec;
         }
     }
 }
